% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{addtime}
\alias{addtime}
\title{Add time in hour to calendar date/time}
\usage{
addtime(datetime, timehour, format = "\%Y-\%m-\%d \%H:\%M", tz = "GMT")
}
\arguments{
\item{datetime}{date/time vector to be computed}

\item{timehour}{Time to be added in hour}

\item{format}{date and time format}

\item{tz}{Time zone (default="GMT")}

\item{add.to.data}{if "yes" result will be appended to dat}

\item{name}{column name of derived variable}
}
\description{
This function allows you to add derived variable (ex: add mean value by ID).
}
\examples{
addtime()
}
\keyword{addtime}
