% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{addvar}
\alias{addvar}
\title{Derived 1 variable and 1 function}
\usage{
addvar(dat, sort, var, fun, add.to.data = "yes", name = NULL)
}
\arguments{
\item{dat}{data frame}

\item{sort}{sort derived variable by (ex:c("ID","SEX"))}

\item{var}{variable to be derived}

\item{fun}{deriving funtion ex:"mean(x)")}

\item{add.to.data}{if "yes" result will be appended to dat}

\item{name}{column name of derived variable}
}
\description{
This function allows you to add derived variable (ex: add mean value by ID).
}
\examples{
addvar()
}
\keyword{addvar}
