% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{addvar2}
\alias{addvar2}
\title{Derived more variables and functions}
\usage{
addvar2(dat, sort, var, fun, rounding = "sigfig(x,3)")
}
\arguments{
\item{dat}{data frame}

\item{sort}{sort derived variable by (ex:c("ID","SEX"))}

\item{var}{variable to be derived}

\item{fun}{deriving funtion ex:c("mean(x)=mean","length(x[is.na(x)])")}
}
\description{
This function allows you to add derived variable (ex: add mean value by ID).
}
\examples{
addvar()
}
\keyword{addvar}
