% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{AUC}
\alias{AUC}
\title{Calculate AUC Using the Trapezoidal Method}
\usage{
AUC(data, time = "TIME", id = "ID", dv = "DV")
}
\arguments{
\item{time}{chronologically ordered time variable present in data}

\item{id}{variable in data defining subject level data}

\item{dv}{dependent variable used to calculate AUC present in data}

\item{data.frame}{containing the data to use for the AUC calculation}
}
\description{
Calculate the area under the curve (AUC) for each subject over the time interval for dv using the trapezoidal rule.
nca.analysis(data,n_lambda=3,id,time,dv,ss,partialAUC="no")
data
}
\examples{
AUC(data, time = 'TIME', id = 'ID', dv = 'DV')
}
\keyword{AUC}
