% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{blk.locf2}
\alias{blk.locf2}
\title{Ben's function}
\usage{
blk.locf2(x, id, na.action = c("fill", "carry.back"), fill = NA)
}
\arguments{
\item{packages}{pre-define packages list.}
}
\description{
internal used
}
\examples{
blk.locf2()
}
\keyword{blk.locf}
