% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{bround}
\alias{bround}
\title{bround Table function}
\usage{
bround(data, var, rtype = "sigfig", dec = 3)
}
\arguments{
\item{data}{data}
}
\description{
bround Table function
}
\examples{
bround()
}
\keyword{bround}
