% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{chclass}
\alias{chclass}
\title{Change variable class}
\usage{
chclass(data, var, class = "char")
}
\arguments{
\item{data}{data}

\item{var}{variable (ex:c("DV","MDV"))}

\item{class}{class ("char" or "num")}
}
\description{
This function allows you to change variable class ("num" or "char").
}
\examples{
chclass()
}
\keyword{chclass}
