% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{conti}
\alias{conti}
\title{conti Table function}
\usage{
conti(input = input, var = var, by = by, round.type = "sigfig",
  digit = 3, quanti)
}
\description{
internal use.
}
\examples{
conti()

}
