% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{cov_fn}
\alias{cov_fn}
\title{Covariate plots function}
\usage{
cov_fn(...)
}
\arguments{
\item{data}{Data frame. Use the codes below as template}
}
\description{
Covariate plots function
}
\keyword{cov_plots()}
