% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{cov_plots}
\alias{cov_plots}
\title{template for Covariate plots}
\usage{
cov_plots(...)
}
\arguments{
\item{data}{Data frame. Use the codes below as template}
}
\description{
template for Covariate plots
}
\keyword{cov_plots()}
