% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{cround}
\alias{cround}
\title{Round up}
\usage{
cround(z, y)
}
\arguments{
\item{z}{Vector or single value to be rounded}

\item{y}{number of significant figure}
}
\description{
This function allows you to round value as in Excel.
}
\examples{
cround()
}
\keyword{rounding}
