% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{cround1}
\alias{cround1}
\title{Rounding as per Excel Internal use}
\usage{
cround1(x, n, asnum = T)
}
\description{
This function allows you to round value as per Excel method.
}
\examples{
cround1()
}
\keyword{cround}
