% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{d010101}
\alias{d010101}
\title{date and time format funtion
Join two datasets and print report of joining procedure}
\usage{
d010101(dattime)
}
\arguments{
\item{dattime}{Date and time data ex: "%d-%b-%y %H:%M:%S"}
}
\description{
date and time format funtion
Join two datasets and print report of joining procedure
}
\examples{
d010101()
}
\keyword{d010101()}
