% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{diftm}
\alias{diftm}
\title{Compute delta using calendar date and time}
\usage{
diftm(tm1, tm2, unit = "hour", form1 = "\%Y-\%m-\%d \%H:\%M",
  form2 = "\%Y-\%m-\%d \%H:\%M", tz = "GMT")
}
\arguments{
\item{tm1}{data frame}

\item{tm2}{subject id}

\item{form1}{date/time format 1}

\item{form2}{date/time format 2}
}
\description{
This function allows you to compute the delta (time1-time2).
}
\examples{
diftm()
}
\keyword{diftm}
