% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{dup1}
\alias{dup1}
\title{Filter unique duplicated row}
\usage{
dup1(data, var, all, select)
}
\arguments{
\item{data}{data}

\item{all}{display all columns (all="all")}

\item{select}{display selected variables only}

\item{data}{data}
}
\description{
This function allows you to filter duplicated rows but only show unique row
}
\examples{
dup1()
}
\keyword{dup1}
