% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{dup2}
\alias{dup2}
\title{Filter all duplicated rows}
\usage{
dup2(data, var, all, select)
}
\arguments{
\item{data}{data}

\item{all}{display all columns (all="all")}

\item{select}{display selected variables only}

\item{data}{data}
}
\description{
This function allows you to filter duplicated rows but only show unique row
}
\examples{
dup2()
}
\keyword{dup2}
