% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{duprow}
\alias{duprow}
\title{Check duplicates}
\usage{
duprow(data, var = NULL, remove = NULL)
}
\arguments{
\item{data}{data}

\item{var}{variable (ex:c("DV","MDV"))}

\item{remove}{if remove="yes", duplicates will be removed)}
}
\description{
This function allows you to check duplicates.
}
\examples{
duprow()
}
\keyword{duprow}
