% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{findiff}
\alias{findiff}
\title{find different values between two datasets}
\usage{
findiff(dat1, dat2)
}
\arguments{
\item{dat1, dat2}{Dataset 1 and 2"}
}
\description{
find different values between two datasets
}
\examples{
findiff()
}
\keyword{findiff()}
