% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{format_time}
\alias{format_time}
\title{Reformat calendar date/time}
\usage{
format_time(dttm, form1, form2 = "\%Y-\%m-\%d \%H:\%M", tz = "GMT")
}
\arguments{
\item{dttm}{original date/time}

\item{form1}{date/time format 1 to be changed}

\item{form2}{new date/time format}

\item{tm2}{subject id}
}
\description{
This function allows you to change the date/time format.
}
\examples{
format_time()
}
\keyword{format_time}
