% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{gof_plots}
\alias{gof_plots}
\title{template for GOF}
\usage{
gof_plots(...)
}
\arguments{
\item{data}{Data frame. Use the codes below as template}
}
\description{
template for GOF
}
\keyword{gof_plots()}
