% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{hl3cpt}
\alias{hl3cpt}
\title{Three compartment micro constants and HL}
\usage{
hl3cpt(data, Cl, Cl2, Cl3, V, V2, V3, output)
}
\arguments{
\item{data}{data}
}
\description{
This function allows you to derive micro constants and HL.
}
\examples{
hl3cpt(pkdat1,"cl","cl2",,"cl3","v","v2","v3","all")
}
\keyword{hl3cpt}
