% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{install.pack}
\alias{install.pack}
\title{install.pack}
\usage{
install.pack(...)
}
\arguments{
\item{packages}{pre-define packages list.}
}
\description{
To install require packages. Use ipak function to install desired packages.
}
\examples{
install.pack()
}
\keyword{install.pack}
