% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhcattab}
\alias{lhcattab}
\title{Funtion for Descriptove Stats of Categorical Covariate}
\usage{
lhcattab(data, var, by)
}
\arguments{
\item{data}{datset or data frame (ex:data=PKdatat)}

\item{var}{List of continuous covariates (ex:c("SEX","RACE"))}

\item{by}{Stratification variable (ex: by="study")}
}
\description{
Descriptove stats of categorical covariates
cat.tab(data,var,by,colby="var",rowby=by)
}
\examples{
cat.tab(data=dat,var=c("SEX","RACE"),by=c("study"),colby="var",rowby=by)
}
\keyword{cat.tab}
