% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhcbind}
\alias{lhcbind}
\title{lhcbind}
\usage{
lhcbind(dat1, dat2)
}
\arguments{
\item{dat1, dat2}{data frames.}
}
\description{
C bind 2 data frames regardless number of row length.
}
\examples{
lhcbind()
}
\keyword{lhcbind}
