% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhcontab}
\alias{lhcontab}
\title{Funtion for Descriptove stats of continuous covariate}
\usage{
lhcontab(dat, var, by, format = "by.var", sumry = 1,
  round.type = "sigfig", digit = 3, quant = 0.9)
}
\arguments{
\item{var}{List of continuous covariates (ex:c("CRCL","WT"))}

\item{by}{Stratification variable (ex: by="study")}

\item{sumry}{Result summary format to be displayed. Type summary.set to see different options of summary sets.}

\item{round.type}{rounding method, sigfig by default. (ex:"round" for rounding)}

\item{digit}{Number of round digits or significant figures}

\item{data}{datset or data frame (ex:data=PKdatat)}

\item{colby}{Specify sorting variable to be displayed vertically. (ex: colby=by or colby="var")}

\item{rowby}{Specidy sorting variable horizontally. If by > 1, set rowby=by.}
}
\description{
Descriptove stats of continuous covariates
}
\examples{
con.tab(data=dat,var=c("WT","crcl","ALT"),by=c("study"),colby="var",rowby=by,sumry=set1)
}
\keyword{con.tab}
