% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhcut}
\alias{lhcut}
\title{Cut values and create category}
\usage{
lhcut(data, var, breaks, labels = "fancy", right = F,
  newvar = "default")
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable to be cut}

\item{breaks}{break points}

\item{labels}{category name. If fancy, the categories will be created according to the break points}

\item{right}{If false, right value will be exclusive
@param newvar vector name of the categorical. If default, the var with suffix"cat" will be used as default name}
}
\description{
Cut values and create category
}
\keyword{lhcut()}
