% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhdoc}
\alias{lhdoc}
\title{lhdoc}
\usage{
lhdoc(toc.level = 4, template = "default", TOC = F,
  style.to.map = "default", logo.certara = F)
}
\arguments{
\item{toc.level}{maximimum toc level}

\item{template}{Word document template could be used for styles. Styles should be mapped in style.to.map. Template is also available at github: to load it, just run  lhtemp() once to download and store the templates in your PC at "c:lhtemplate. Note that the templates and logo are also used in xptdef package.}

\item{TOC}{Set to F if no TOC wanted}

\item{style.to.map}{Map the styles in template to be used. Ex: mypar is for footnote (font size)}

\item{t}{template name see at the end of lhtext function.}
}
\description{
Create doc for word document using this fonction or you can do it manually.
Type lhtext and copy the template to R workspace and start writing.
}
\keyword{lhdoc}
