% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhfactor}
\alias{lhfactor}
\title{Change factor level of a variable using matched level of another variable in the same dataset}
\usage{
lhfactor(data = test, leader = "AGEcat", follower = "catt")
}
\arguments{
\item{data}{Data frame}

\item{leader}{lead Variable to be used for factor level of the follower variable}

\item{follower}{follower Variable}
}
\description{
Change factor level of a variable using matched level of another variable in the same dataset
}
\keyword{lhfactor()}
