% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhloess}
\alias{lhloess}
\title{lhloess}
\usage{
lhloess(data, x, y, by, span = 1)
}
\arguments{
\item{data}{data.}

\item{x}{Independent variable}

\item{y}{Dependent variable}

\item{by}{Sort by. Only one sorting variabele is accepted. If more than 1 variables, create a unique sorting using paste(var1,var2,etc)}

\item{span}{LOESS stiffness}
}
\description{
Compute the LOESS data for ploting.
}
\examples{
lhloess()
}
\keyword{lhloess}
