% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhlong}
\alias{lhlong}
\title{Reshape long}
\usage{
lhlong(data, long.vector)
}
\arguments{
\item{data}{Dataset}

\item{long.vector}{List of vector to be melted}
}
\description{
Reshape long
}
\examples{
lhlong()
}
\keyword{lhlong()}
