% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhlook}
\alias{lhlook}
\title{Look for keyword across dataset}
\usage{
lhlook(var, file.or.path, fpattern = NULL, filename = NULL)
}
\arguments{
\item{var}{Variable to be cut}

\item{file.or.path}{File path ("xxx/zzz/") or data frame (c("lb","dm"))}

\item{fpattern}{File extention: ex. "csv", all csv file in folder will be used}

\item{filename}{specify full file name. set fpattern to Null}

\item{data}{Data frame}
}
\description{
Look for keyword across dataset
}
\keyword{lhlook()}
