% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhmerge_wide_data}
\alias{lhmerge_wide_data}
\title{merge wide data frames}
\usage{
lhmerge_wide_data(dat1, dat2, by1, by2)
}
\arguments{
\item{dat1, dat2}{Data frame 1 and 2 with long vectors and values. Note: no duplicated sorting vector allowed}

\item{by1, by2}{Vectors containing matched values from each data frames.}
}
\description{
merge wide data frames
}
\keyword{lhmerge_long_data()}
