% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhmutate}
\alias{lhmutate}
\title{mutate variable names}
\usage{
lhmutate(data, mutate)
}
\arguments{
\item{data}{Data frame 1 and 2 with long vectors and values. Note: no duplicated sorting vector allowed}

\item{mutate}{Vector to be mutated ex. "xxx=yyy" for renaming xxx as yyy}
}
\description{
mutate variable names
}
\keyword{lhmutate()}
