% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhorder}
\alias{lhorder}
\title{lhorder funtion}
\usage{
lhorder(dat, var)
}
\arguments{
\item{dat}{Datframe}

\item{var}{Order by variables. ex: ":Trt,:Agegr"}
}
\description{
Make simple table. Use data frame created by addvar2
}
\examples{
lhorder()
}
\keyword{lhorder()}
