% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhrelabel}
\alias{lhrelabel}
\title{Relabel variable and factor}
\usage{
lhrelabel(data, var, kfactor = T)
}
\arguments{
\item{data}{Data frame 1 and 2 with long vectors and values. Note: no duplicated sorting vector allowed}

\item{var}{c("varname","label1=label2") "xxx=yyy" for renaming xxx to yyy}
}
\description{
Relabel variable and factor
}
\keyword{lhrelabel()}
