% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhtab}
\alias{lhtab}
\title{lhtab funtion}
\usage{
lhtab(data, vh, value, ord = NULL, save.name = NULL, output = "csv")
}
\arguments{
\item{data}{Datframe}

\item{vh}{Vertical and horizontal headers. ex: "Trt+Agegr~Param"}

\item{value}{Values. Example: c("mean","SD","Mean (CV)")}

\item{ord}{Order variables. ex: ":Trt,:Agegr"}

\item{save.name}{Save table as word document. Enter the file name: ex "test.docx"}

\item{output}{output="csv" for csv output, else output will be in FlexTable format}
}
\description{
Make simple table. Use data frame created by addvar2
}
\examples{
lhtab()
}
\keyword{lhtab()}
