% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhtemplate}
\alias{lhtemplate}
\title{lhtemplate}
\usage{
lhtemplate()
}
\arguments{
\item{repo}{leonpheng}

\item{user}{logo.style}
}
\description{
To install require packages. Use ipak function to install desired packages.
}
\examples{
lhtemplate()
}
\keyword{lhtemplate}
