% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhtext}
\alias{lhtext}
\title{lhtext}
\usage{
lhtext(t, toc.level = 3,
  template = "C:/Users/User/Desktop/packdev/styleapp.docx", TOC = T,
  style.to.map = c("Heading1", "Heading2", "Heading2", "fnt"),
  working.folder = NULL, logo.certara = "none")
}
\arguments{
\item{t}{list of items. see example.}

\item{toc.level}{maximimum toc level}

\item{template}{Word document template could be used for styles. Styles should be mapped in style.to.map. Template is also available at github: to load it, just run  lhtemp() once to download and store the templates in your PC at "c:lhtemplate. Note that the templates and logo are also used in xptdef package.}

\item{TOC}{Set to F if no TOC wanted}

\item{style.to.map}{Map the styles in template to be used. Ex: mypar is for footnote (font size)}
}
\description{
Simple way to create word document using loop from list t.
Type lhtext and copy the template to R workspace and start writing.
}
\examples{
TEMPLATE
t<-NULL
t[[1]]<-c("lev1","Ceci est un test. lev1= level 1 title")
t[[2]]<-c("txtn","bold","Add bold paragragh with page break","pageb") #normal simple text
t[[3]]<-c("lev2","Title for level 2. Below is for creating table")
tab0<-data.frame(x=c("MONDAY","TUESDAY"),y=c("EAT SOMETHING","DON'T EAT"))
t[[4]]<-c("tab","FlexTable(tab0)")
t[[5]]<-c("fnt","This is foot not with font size of 9 by default","pageb")
t[[6]]<-c("lev2","This a test Plot- Used plot. ggplot could be inserted")
t[[7]]<-c("fig","plot(1,1)",5,5,"pageb")
tab1<-read.csv("1-OH-Midazolam1.25+1.25mg.csv")
t[[8]]<-c("lev2","Another Table from folder")
t[[9]]<-c("tab","FlexTable(tab1)","pageb")
t[[10]]<-c("lev2","This a test Image imported from folder and inserted in doc")
t[[11]]<-c("ima","hydroxy-auc0.6-v-age-1.png",7.4,6)
t[[12]]<-c("txtc","This is complex text editor is ::b:i"," X","2::e:u" ,"bold::b:i:s","fwb::e","this::i","fsti::b:i") #formatted text.
tips for symbol, example: ++a++, ++inf++, ++n++, etc.. then do search and replace in Word
e=superscrip, s=subscript, b=bold, i=italic, ::= start code, := additional code
lhtext(t,toc.level=3,template="style.docx",TOC=F)
to write doc: writeDoc(doc,"name.docx")
lhtext()
}
\keyword{lhtext}
