% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhvpc_stat}
\alias{lhvpc_stat}
\title{Compute VPC stats}
\usage{
lhvpc_stat(obs.data = obs, sim.data = sim, bin = "bin",
  prob = c(0.025, 0.5, 0.95), sort = NULL, dv = "DV", tad = "TAD",
  rtime = "IVAR", blq = NULL, replicate = "REPLICATE",
  pred.corr = NULL)
}
\arguments{
\item{obs.data}{Observed data. Type vpc_plots forbasic R code to be used with the output}

\item{sim.data}{Sim data}

\item{bin}{Binning. This should be created in both datasets}

\item{prob}{Quantile}

\item{sort}{Sorting as vector}

\item{dv}{Concentration or response.Same name in both datasets}

\item{tad}{Time after dose. In obs.data}

\item{rtime}{Time after first dose. In obs.data}

\item{blq}{LLOQ. Could be single value or vector. If vector, both datasets should have the same variable name. This will compute the % of BLQ for each bin.}

\item{replicate}{Replicate in sim.data only}

\item{pred.corr}{Pred-corrected. ex: C("PRED","lin") use the column PRED and corrected in linear scale or log for log scale}
}
\description{
Compute VPC stats
}
\keyword{lhvpc_stat()}
