% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhwide}
\alias{lhwide}
\title{Reshape wide}
\usage{
lhwide(data, wide.data, wide.vector)
}
\arguments{
\item{data}{Dataset}

\item{wide.data}{Name of vector containing data to be dcasted
@param wide.vector Name of vector to be reshape as heading}

\item{data}{Dataset}
}
\description{
Reshape wide
}
\examples{
lhwide()
}
\keyword{lhwide()}
