% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{loadpack}
\alias{loadpack}
\title{load.pack1}
\usage{
loadpack(...)
}
\arguments{
\item{packages}{pre-define packages list.}
}
\description{
To install require packages. Use ipak function to install desired packages.
}
\examples{
load.pack1()
}
\keyword{load.pack1}
