% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{locf2}
\alias{locf2}
\title{LOCF and LOCB}
\usage{
locf2(data, by, var, locb = T)
}
\arguments{
\item{data}{data}

\item{by}{sort variable}

\item{var}{variable to locf}

\item{locb}{carry backward}
}
\description{
LOCF LOCB function
}
\examples{
locf2()
}
\keyword{locb2}
