% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{m6}
\alias{m6}
\title{BLQ M6 Method}
\usage{
m6(data, id, evid, mdv, blq.flag, time, dv, lloq)
}
\arguments{
\item{data}{data frame}

\item{id}{ID vector}

\item{evid}{EVID vector}
}
\description{
This function allows you to create data with BLQ using M6 method.
}
\examples{
tad_addl()
}
\keyword{m6}
