% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{nca.cal}
\alias{nca.cal}
\title{Derive Common NCA parameters using single and multiple profiles}
\usage{
nca.cal(data, n_lambda = 3, id = "id", time = "time", dv = "dv",
  sort.by = NULL, ss.variable = "ss", sd_label_in_ss = NULL,
  ss_label_in_ss = NULL, tau = NULL, dose = NULL,
  partialAUC = NULL, partialConc = NULL)
}
\arguments{
\item{data}{datset or data frame (ex:data=PKdatat)}

\item{n_lambda}{number of points for estimating the Lambda}

\item{id}{unique subject identifier}

\item{time}{Sampling time after dose (TAD)}

\item{dv}{Concentration}

\item{tau}{Dose interval. As vector or numeric.}

\item{dose}{Dose administered. As vector or numeric. Required to estimate CL and Vss}

\item{partialAUC}{Time interval for partial AUC. Ex: c(0,6,0,12,6,12) for AUC0-6, AUC0-12 and AUC6-12}

\item{partialConc}{Concentration at particular time (Ex:c(1,4) for concentration after 1 and 4 h)}

\item{dosing.regimen}{PK profile profile flag (ex: single, sd, md ,ss).As vector or numeric. Note that only one prefile per dose is to be analyzed.}

\item{label.for.sd}{Label to identify sd profile used in dosing.regimen. Need to compute the effective half-life}

\item{label.for.ss}{Label to identify ss profile used in dosing.regimen. Need to compute the effective half-life}
}
\description{
nca.cal()
}
\examples{
p<-nca.cal(data=data, n_lambda = 3, id = "id", time = "time", dv = "dv", 
tau="ii",dose="dose",dosing.regimen = "ss",label.for.sd="single", partialAUC = NULL, partialConc = NULL)

}
\keyword{nca.cal}
