% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{nmiss}
\alias{nmiss}
\title{nmiss Table function
internal use.}
\usage{
nmiss(x)
}
\arguments{
\item{x}{data}
}
\description{
nmiss Table function
internal use.
}
\examples{
nmiss()
}
\keyword{nmiss}
