% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{nodup}
\alias{nodup}
\title{No duplicate}
\usage{
nodup(data, var, all, item)
}
\arguments{
\item{data}{data}

\item{var}{variable (ex:c("DV","MDV"))}

\item{all}{if all="all", all columns in data will be kept (ex:all=c("ID","DV"))}
}
\description{
This function allows you to remove duplicates.
}
\examples{
nodup()
}
\keyword{nodup}
