% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{rt2tad}
\alias{rt2tad}
\title{Derive TAD from RTIME}
\usage{
rt2tad(data, id, time, evid)
}
\arguments{
\item{data}{data frame}

\item{id}{subject id}

\item{time}{time after first dose}

\item{evid}{evid}
}
\description{
This function allows you to derive Time after dose from Time after first dose.
}
\examples{
rt2tad()
}
\keyword{rt2tad}
