% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{sigfig}
\alias{sigfig}
\title{Significant figure}
\usage{
sigfig(z, y)
}
\arguments{
\item{z}{Vector or single value to be rounded}

\item{y}{number of significant figure}
}
\description{
This function allows you to round value in significant figure.
}
\examples{
sigfig()
}
\keyword{sigfig}
