% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{sigfig1}
\alias{sigfig1}
\title{Significant figure}
\usage{
sigfig1(x, y)
}
\description{
This function allows you to round value in significant figure.
}
\examples{
sigfig1()
}
\keyword{sigfig}
