% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{stackvar}
\alias{stackvar}
\title{Combine variables in the same column}
\usage{
stackvar(data, combine.var = c("xxx", "variable"))
}
\arguments{
\item{data}{Data frame 1 and 2 with long vectors and values. Note: no duplicated sorting vector allowed}

\item{combine.var}{Variable name, c(var1,var2). var1 will be stacked over var 2}
}
\description{
Combine variables in the same column
}
\keyword{stackvar()}
