% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{tab}
\alias{tab}
\title{Quick table}
\usage{
tab(data, var1, var2)
}
\arguments{
\item{data}{original date/time}

\item{var1}{var1}

\item{var2}{var2}
}
\description{
This function allows you to create quick and dirty table.
}
\examples{
tab()
}
\keyword{tab}
