% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{tad_addl}
\alias{tad_addl}
\title{TAD from ADDL}
\usage{
tad_addl(data = df, id = "usubjid", ii = "ii", addl = "addl",
  rtime = "rtime", evid = "evid", dose.expand = "yes")
}
\arguments{
\item{data}{data frame}

\item{id}{ID vector}

\item{ii}{dose interval vector}

\item{addl}{additional dose vector}

\item{rtime}{relative time after first dose vector}

\item{evid}{EVID vector}

\item{dose.expand}{If "yes", all dosing rows in ADDL will be outputed}

\item{dose}{amount adminstered (ex: AMT) vector}
}
\description{
This function allows you to derive time after dose from ADDL.
}
\examples{
tad_addl()
}
\keyword{tad}
