% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{txt}
\alias{txt}
\title{txt funtion}
\usage{
txt(c)
}
\arguments{
\item{c}{text. Example: c("Concentration mg L","-1::s"," AUC::u"," Delta::i","moles::e"). s=subscript, u=underline, Delta= capital greek delta letter, i= italic, e=superscript}
}
\description{
Clone expression function for adding special formats and symbol to plots.
}
\examples{
txt()
}
\keyword{txt()}
