% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{vpc_plots}
\alias{vpc_plots}
\title{template for VPC}
\usage{
vpc_plots(...)
}
\arguments{
\item{data}{Data frame Use the codes below as template}
}
\description{
template for VPC
}
\keyword{vpc_plots()}
