% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{lhmerge_long_data}
\alias{lhmerge_long_data}
\title{merge and make long to wide data frame}
\usage{
lhmerge_long_data(dat1, dat2, long.vector1, long.vector2, value1, value2)
}
\arguments{
\item{dat1, dat2}{Data frame 1 and 2 with long vectors and values. Note: no duplicated sorting vector allowed}

\item{long.vector}{Vector containing wide variable names}

\item{value}{Column containing values associated with each wide variable.}
}
\description{
merge and make long to wide data frame
}
\keyword{lhmerge_long_data()}
