% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{nca.cal}
\alias{nca.cal}
\title{Derive Common NCA parameters using single profile}
\usage{
nca.cal(data, n_lambda = 3, id = "id", time = "time", dv = "dv",
  ss = "ss", partialAUC = NULL, partialConc = NULL)
}
\arguments{
\item{data}{datset or data frame (ex:data=PKdatat)}

\item{n_lambda}{number of points for estimating the Lambda}

\item{id}{unique subject identifier}

\item{time}{Sampling time after dose (TAD)}

\item{dv}{Concentration}

\item{ss}{Steady state flags. Note that ss will be used only for identification purpose only.}

\item{partialAUC}{Time interval for partial AUC. Ex: c(0,6,0,12,6,12) for AUC0-6, AUC0-12 and AUC6-12}

\item{partialConc}{Concentration at particular time (Ex:c(1,4) for concentration after 1 and 4 h)}
}
\description{
nca.analysis(data,n_lambda=3,id,time,dv,ss,partialAUC="no")
}
\examples{
nca.analysis(data=dat,id="NMID",by=c("study"))
}
\keyword{nca.analysis}
