% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectedfunctions.R
\name{vpc_plots}
\alias{vpc_plots}
\title{Look for keyword across dataset}
\usage{
vpc_plots(...)
}
\arguments{
\item{data}{Data frame}
}
\description{
Look for keyword across dataset
}
\keyword{vpc_plots()}
